I2-60 ref: Refactor Payment component
