+++ src/widgets/Payments/ui/Payments.module.scss
